+++ test/e2e/tests/hardware-wallets/ledger/ledger-forget-device.spec.ts
@@ -1,6 +1,6 @@
 import { Suite } from 'mocha';
 import { Browser } from 'selenium-webdriver';
-import FixtureBuilder from '../../../fixture-builder';
+import FixtureBuilder from '../../../fixtures/fixture-builder';
 import { withFixtures } from '../../../helpers';
 import HomePage from '../../../page-objects/pages/home/homepage';
 import { Driver } from '../../../webdriver/driver';