+++ ui/components/component-library/select-wrapper/select-wrapper.stories.tsx
@@ -172,8 +172,9 @@ UncontrolledValue.args = {
 };
 
 export const UseSelectContext: StoryFn<typeof SelectWrapper> = (args) => {
-  // Note that the SelectContext is being used inside a component, because the SelectContext needs to be called within the SelectWrapper component and not before
   const CustomClose = () => {
+    // Note that the SelectContext is being used inside a component, because the SelectContext needs to be called within the SelectWrapper component and not before
+    // eslint-disable-next-line react-compiler/react-compiler
     const { toggleUncontrolledOpen } = useSelectContext();
 
     return (